requirements.txt
google-api-core
google-auth>=2.15.0,<3.0.0
google-auth-oauthlib>=0.8.0
google-cloud-bigquery==3.4.1
google-cloud-bigquery-storage==2.18.0
google-cloud-core>=2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
apache-airflow==2.6.3
apache-airflow-providers-ftp==3.5.1
apache-airflow-providers-google==10.2.0
apache-airflow-providers-http==4.5.1
apache-airflow-providers-imap==3.3.1
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-sqlite==3.4.3
xmltodict